new fixes

+++ MvcApplication1/Controllers/ValuesController.cs
@@ -2,6 +2,7 @@
 using DataAccess.Models;
 using System;
 using System.Collections.Generic;
+using System.Diagnostics;
 using System.Linq;
 using System.Net;
 using System.Net.Http;
@@ -11,6 +12,8 @@ namespace MvcApplication1.Controllers
 {
     public class ValuesController : ApiController
     {
+        HubClient c = new HubClient(Constants.Host, Clients.Server);
+        object locking = new object();
         //GET api/values/GetFileInfo
         [HttpGet]
         [ActionNameAttribute("GetFileInfo")]
@@ -20,9 +23,7 @@ namespace MvcApplication1.Controllers
             {
                 var f = context.Files.FirstOrDefault();
                 if (f != null)
-                {
-                    context.Files.Remove(f);
-                    context.SaveChanges();
+                {                   
                     return f;
                 }
                 return null;
@@ -37,39 +38,47 @@ namespace MvcApplication1.Controllers
             using (DataBaseContext context = new DataBaseContext())
             {
                 context.Files.Add(info);
-                context.SaveChanges();
-                this.SendMsg(Clients.Downloader, Messages.FileInfoAvailable);
+                context.SaveChanges();                
+                c.SendMessage(new MsgData { From = Clients.Server, To = Clients.Downloader, Message = Messages.FileInfoAvailable });                
             }
         }
 
-        // POST api/values/RemoveFileInfo/{id}
-        [ActionNameAttribute("RemoveFileInfo")]
+        // DELTE api/values/RemoveFileInfo/{id}
+        [HttpDelete]
+        [ActionNameAttribute("RemoveFileInfo")]        
         public void DeleteFileInfo(long id)
         {
             using (DataBaseContext context = new DataBaseContext())
             {
                 var file = context.Files.FirstOrDefault(p => p.Id == id);
-                context.Files.Remove(file);
-                context.SaveChanges();
+                if (file != null)
+                {
+                    context.Files.Remove(file);
+                    context.SaveChanges();
+                }
             }
         }
 
 
         // GET api/values/AnyPart
         [HttpGet]
-        [ActionNameAttribute("AnyPart")]
-        public FilePart AnyPart()
+        [ActionNameAttribute("GetPart")]
+        public FilePart GetPart(long id)
         {
             using (DataBaseContext context = new DataBaseContext())
             {
-                var part = context.Parts.FirstOrDefault();
+                var part = context.Parts.FirstOrDefault(p => !p.IsTaken && p.Id == id);
                 if (part != null)
-                {                    
+                {
+                    Debug.WriteLine("{0} part taking", part.Part);
+
                     part.IsTaken = true;
                     context.SaveChanges();
 
-                    if (context.Parts.Count(t => !t.IsTaken) < 3)
-                        this.SendMsg(Clients.Uploader, Messages.ContinueUploading);
+                    //if (context.Parts.Count(t => !t.IsTaken) < 2)
+                    //    c.SendMessage(new MsgData { From = Clients.Server, To = Clients.Uploader, Message = Messages.ContinueUploading });
+                    //else
+                    //    c.SendMessage(new MsgData { From = Clients.Server, To = Clients.Uploader, Message = Messages.PauseUploading });
 
                     return part;
                 }
@@ -80,42 +89,52 @@ namespace MvcApplication1.Controllers
 
         // POST api/values/AddPart
         [HttpPost]
-        [ActionNameAttribute("AddPart")]
+        [ActionNameAttribute("AddPart")]        
         public void Post(FilePart newPart)
         {
             using (DataBaseContext context = new DataBaseContext())
             {
                 context.Parts.Add(newPart);
-                context.SaveChanges();
-
-                if (context.Parts.Count(t => !t.IsTaken) > 8)
-                    this.SendMsg(Clients.Uploader, Messages.PauseUploading);
-                else
-                    this.SendMsg(Clients.Uploader, Messages.ContinueUploading);
-
-                if (context.Files.Any())
-                    this.SendMsg(Clients.Downloader, Messages.FileInfoAvailable);
-
-                this.SendMsg(Clients.Downloader, Messages.DownloadAvailable);
+                Debug.WriteLine(string.Format("{0} parts on server", context.Parts.Count()));
+                context.SaveChanges();                
+                c.SendMessage(new MsgData { From = Clients.Server, To = Clients.Downloader, Message = string.Format("{0}{1}", newPart.Id, Messages.DownloadAvailable) });
             }
         }
 
         // POST api/values/RemovePart/{id}
+        [HttpDelete]
         [ActionNameAttribute("RemovePart")]
         public void Delete(long id)
         {
             using (DataBaseContext context = new DataBaseContext())
             {
                 var part = context.Parts.FirstOrDefault(p => p.Id == id);
-                context.Parts.Remove(part);
-                context.SaveChanges();
+                if (part != null)
+                {                    
+                    context.Parts.Remove(part);
+                    context.SaveChanges();
+                    Debug.WriteLine("{0} part removed", part.Part);
+                }
             }
         }
 
-        private void SendMsg(string to, string msg)
+        [HttpDelete]
+        [ActionNameAttribute("ClearDb")]
+        public void ClearDb()
         {
-            HubClient c = new HubClient(Constants.Host, Clients.Server);
-            c.SendMessage(new MsgData { From = Clients.Server, To = to, Message = msg });
+            using (DataBaseContext context = new DataBaseContext())
+            {
+                var objCtx = ((System.Data.Entity.Infrastructure.IObjectContextAdapter)context).ObjectContext;
+                objCtx.ExecuteStoreCommand("TRUNCATE TABLE [Parts]");
+                objCtx.ExecuteStoreCommand("TRUNCATE TABLE [Files]");
+            }
         }
+
+
+        //private void SendMsg(string to, string msg)
+        //{
+            
+            
+        //}
     }
 }
